requirements.txt
crewai
langchain-community
langchain-openai
langchain-core
pypdf
duckduckgo-search
yfinance
pandas
python-dotenv
